Dev zz

+++ src/main/java/com/vo/ModelVO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class ModelVO {
     public Integer modelId;
-    private Long manId;
+    private Integer manId;
     private String modelName;
     private String year;
     private Integer seatNum;
